backend/sandbox-gateway/requirements.txt
docker>=7.0.0,<8.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets==10.4
aiohttp==3.9.3
python-dotenv==1.0.1
pydantic-settings>=2.5.2
anyio>=4.5.0
httpx>=0.27.0